Fizermos muitas coisas, tela de pesquisar funfando e conectada no banco e é.

+++ src/bean/PedidosProdutos.java
@@ -2,7 +2,6 @@ package bean;
 // Generated 15/09/2025 10:21:54 by Hibernate Tools 4.3.1
 
 
-import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -27,12 +26,12 @@ public class PedidosProdutos  implements java.io.Serializable {
      private Pedidos pedidos;
      private Produtos produtos;
      private Integer quantidade;
-     private BigDecimal valorUnitario;
+     private Double valorUnitario;
 
     public PedidosProdutos() {
     }
 
-    public PedidosProdutos(Pedidos pedidos, Produtos produtos, Integer quantidade, BigDecimal valorUnitario) {
+    public PedidosProdutos(Pedidos pedidos, Produtos produtos, Integer quantidade, Double valorUnitario) {
        this.pedidos = pedidos;
        this.produtos = produtos;
        this.quantidade = quantidade;
@@ -83,11 +82,11 @@ public class PedidosProdutos  implements java.io.Serializable {
 
     
     @Column(name="valorUnitario", precision=10)
-    public BigDecimal getValorUnitario() {
+    public Double getValorUnitario() {
         return this.valorUnitario;
     }
     
-    public void setValorUnitario(BigDecimal valorUnitario) {
+    public void setValorUnitario(Double valorUnitario) {
         this.valorUnitario = valorUnitario;
     }
